+++ GameBoy/GB/Display/IDisplay.cs
@@ -2,7 +2,7 @@
 {
     public interface IDisplay
     {
-        public void DrawLine(int y, byte[] line);
+        public void DrawLine(int y, uint[] line);
         public void Render();
     }
 }
